requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
python-telegram-bot==20.7
matplotlib==3.8.2